+++ prisma/migrations/20260907171345_add_per_doc_branding/migration.sql
@@ -1,0 +1,9 @@
+-- AlterTable
+ALTER TABLE "BrandingSetting" ADD COLUMN     "purchaseFooter" TEXT,
+ADD COLUMN     "purchaseReturnFooter" TEXT,
+ADD COLUMN     "purchaseReturnTerms" TEXT,
+ADD COLUMN     "purchaseTerms" TEXT,
+ADD COLUMN     "saleFooter" TEXT,
+ADD COLUMN     "saleTerms" TEXT,
+ADD COLUMN     "salesReturnFooter" TEXT,
+ADD COLUMN     "salesReturnTerms" TEXT;